bug fixes
bug fixes

+++ CCDS/SaveAllPNG.hpp
@@ -15,7 +15,7 @@
 #include <cmath>
 
 using namespace blitz;
-void SaveAllPNG(int Visual_range_x, int Visual_range_y, Array<float,2> cell_array, int H, int &T, float alpha, float beta,double deltah)
+void SaveAllPNG(int Visual_range_x, int Visual_range_y, Array<double,2> cell_array, int H, int &T, double alpha, double beta,double deltah)
 {
     int DELTA=10;
     if (H%DELTA==0)
