+++ CCDS/SavePNGHR.hpp
@@ -16,7 +16,7 @@
 #include <cmath>
 
 using namespace blitz;
-void SavePNGHR(int Visual_range_x, int Visual_range_y, Array<float,2> cell_array, int H, int &T, float alpha, float beta,double deltah)
+void SavePNGHR(int Visual_range_x, int Visual_range_y, Array<double,2> cell_array, int H, int &T, double alpha, double beta,double deltah)
 {
     
     //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
