+++ CCDS/SavePNGSingleCell.hpp
@@ -16,7 +16,7 @@
 #include <cmath>
 
 using namespace blitz;
-void SavePNGSingleCell(int Visual_range_x, int Visual_range_y, int &T, float alpha, float beta, Array<float,2> cell_array)
+void SavePNGSingleCell(int Visual_range_x, int Visual_range_y, int &T, double alpha, double beta, Array<double,2> cell_array)
 {
     /////////////////////////////////////////////////PNG//////////////////////////////////////////////////////////////
     char filedir4 [100] = {'\0'};
