+++ CCDS/sortRowReverse.hpp
@@ -28,7 +28,7 @@ int cmp2( const void *a ,const void *B)
 
 using namespace blitz;
 using namespace std;
-void sortRowReverse(Array<float, 2> &cell_array, Array<float, 2> cell_array1, int Col, int colnum_to_sort,int threads)
+void sortRowReverse(Array<double, 2> &cell_array, Array<double, 2> cell_array1, int Col, int colnum_to_sort,int threads)
 {
     switch (threads)
     {
